Update mbdev-headlines-3_block.utl
Added new custom property css_classes. This allows you to add whatever custom classes you like to the block. This replicates the functionality of the one built into Blox except that it'll respect when the block has no assets. That means, if you apply a padding class to a block that sometimes appears devoid of assets, you won't have extra white space hanging around on the page.